requirements312.txt
altgraph==0.17.4
coloredlogs==15.0.1
flatbuffers==25.9.23
humanfriendly==10.0
ml_dtypes==0.5.3
mpmath==1.3.0
natsort==8.4.0
numpy==2.2.6
onnx==1.19.1
onnxruntime==1.23.2
opencv-python==4.12.0.88
packaging==25.0
pefile==2023.2.7
pillow==12.0.0
protobuf==6.33.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
pyreadline3==3.5.4
pywin32-ctypes==0.2.3
QtAwesome==1.4.0
QtPy==2.4.3
setuptools==80.9.0
sympy==1.14.0
typing_extensions==4.15.0
wheel==0.45.1
